+++ .gitignore
@@ -1,1 +1,2 @@
-node_modules
+frontend/node_modules
+backend/node_modules
